% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{add_summaries}
\alias{add_summaries}
\title{Add summary features to a `Biobase::ExpressionSet`}
\usage{
add_summaries(eset, group_ind, fun = sum)
}
\arguments{
\item{eset}{A `Biobase::ExpressionSet`.}

\item{group_ind}{The name of the column of dData(eset) to use to find
values to aggregate.}

\item{fun}{Aggregatation function to apply to each column. By default `sum`.}
}
\value{
A `Biobase::ExpressionSet`.
}
\description{
Add summary features to a `Biobase::ExpressionSet`
}
