% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{aggregate_eset}
\alias{aggregate_eset}
\title{Aggregate ExpressionSet values according to an indicator column}
\usage{
aggregate_eset(eset, ind_dupes, fun)
}
\arguments{
\item{eset}{A `Biobase::ExpressionSet`.}

\item{ind_dupes}{The name of the column of pData(eset) to use to find
duplicate values to aggregate.}

\item{fun}{Aggregatation function. Should be able to handle both numeric and
character vectors.}
}
\value{
A `Biobase::ExpressionSet`.
}
\description{
Aggregate ExpressionSet values according to an indicator column
}
