% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{an}
\alias{an}
\title{Annotations for Biocrates IDQ p180.}
\format{An object of class \code{data.frame} with 188 rows and 9 columns.}
\source{
\url{http://www.smpdb.ca/}
}
\usage{
an
}
\description{
Annotation informations based on the SMPDB.
}
\keyword{datasets}
