% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{batches}
\alias{batches}
\title{Get batch information from plate production number.}
\usage{
batches(eset, batch_ind = "Plate Production No.")
}
\arguments{
\item{eset}{A Biobase::ExpressionSet.}

\item{batch_ind}{The name of the column containing the information to use for
determining the batches. Defaults to "Plate Production No.).}
}
\value{
Batches as factor.
}
\description{
For large experiments with many samples, more than one IDQ Kit is often
  necessary. However, the results are often combined into one file. This
  function extracts the batch inforlation using the plate production number
  or any desired column. NAs are considered one batch.
}
