% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{cvs_from_qcs}
\alias{cvs_from_qcs}
\title{Calculate CVs using QCs.}
\usage{
cvs_from_qcs(
  eset,
  ind_qcs = "Sample Identification",
  by_batch = T,
  batch_ind = "Plate Production No."
)
}
\arguments{
\item{eset}{A Biobase::ExpressionSet.}

\item{ind_qcs}{Name of the column to use to identify QCs (any sample
measured at least twice)}

\item{by_batch}{Logical telling if CVs should be calculated by batch.
If calculations are performed before removal of batch effect this could
result in higher CVs than expected.}

\item{batch_ind}{The name of the column from where to derive batch
information.}
}
\value{
A data.frame with CVs by QC sample, metabolite and optionally batch.
}
\description{
Calculate CVs using QCs.
}
