% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{LOD_matrix}
\alias{LOD_matrix}
\title{Get the matrix of LOD values for each metabolite with respect to batches.}
\usage{
LOD_matrix(eset)
}
\arguments{
\item{eset}{A Biobase::ExpressionSet.}
}
\value{
A matrix of LOD values of the same size as `exprs(eset)`.
}
\description{
Get the matrix of LOD values for each metabolite with respect to batches.
}
