% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_utils.R
\name{log_transform}
\alias{log_transform}
\title{Log transfrom the data.}
\usage{
log_transform(eset, log_base = 2)
}
\arguments{
\item{eset}{A Biobase::ExpressionSet.}

\item{log_base}{The log base to use. Typically 2.}
}
\value{
A `Biobase::ExpressionSet()`.
}
\description{
This log transformation is robust again negative, zero and missing values.
}
