% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p180_annot}
\alias{p180_annot}
\title{Annotations for Biocrates IDQ p180.}
\format{An object of class \code{data.frame} with 188 rows and 9 columns.}
\source{
\url{http://www.smpdb.ca/}
}
\usage{
p180_annot
}
\description{
Annotation informations based on the SMPDB. Some metabolites
  determined by FIA correspond to multiple isomeres. In these cases either
  the first best match in the HMDB database was taken, or no database ID was
  provided.
}
\keyword{datasets}
