% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pathway_list}
\alias{pathway_list}
\title{Provides a list of pathways associated to each feature by tidying the
  annotations of an eset.}
\usage{
pathway_list(met_df, met_var, path_var, path_desc = NULL, path_sep = "; ")
}
\arguments{
\item{met_df}{A `data.frame` of feature informations, typically `fData(eset)`.}

\item{met_var}{The name of the variable containing the feature IDs. If equal
to "rownames", the row names are used as IDs.}

\item{path_var}{The name of the variable containing the pathway IDs.}

\item{path_desc}{The name of the variable containing the pathway descriptors
(optional).}

\item{path_sep}{Character to use to split up the different pathways and
descriptions associated with each feature.}
}
\value{
A `data.frame` with one row per pathway associated with each feature.
}
\description{
Provides a list of pathways associated to each feature by tidying the
  annotations of an eset.
}
