% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_densities}
\alias{plot_densities}
\title{Plot the density of all features.}
\usage{
plot_densities(eset, alpha = 0.15, plot = T)
}
\arguments{
\item{eset}{A Biobase::ExpressionSet.}

\item{alpha}{Level of transparency to use for the density curves.}

\item{plot}{Logical. Should the plot be plotted right away?}
}
\value{
An invisible `ggplot()`.
}
\description{
Plot the density of all features.
}
