% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_layout}
\alias{plot_layout}
\title{Plot the assay plate layout.}
\usage{
plot_layout(
  eset,
  highlight,
  fill = "Sample Type",
  well_col = "Well Position",
  batch_ind = "Plate Production No."
)
}
\arguments{
\item{eset}{A Biobase::ExpressionSet.}

\item{highlight}{An optional expression used to highlight certain samples.}

\item{fill}{Name of the column used to distinguish samples and wells.}

\item{well_col}{Name of the column containing the well position.}

\item{batch_ind}{The name of the column containing the information to use for
determining the batches. Defaults to "Plate Production No.).}
}
\value{
An invisible `ggplot()`.
}
\description{
Plot the assay plate layout.
}
