% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_pca}
\alias{plot_pca}
\title{Produce a PCA plot.}
\usage{
plot_pca(
  eset,
  axes = 1:2,
  scaling = T,
  group,
  ellipse = F,
  ellipse_prob = 0.68,
  annotations
)
}
\arguments{
\item{eset}{A Biobase::ExpressionSet.}

\item{axes}{Numeric vector of length two specifying which components to draw.}

\item{scaling}{Logical specifying if scaling should be performed, or function
to use to perform scaling. Defaults to TRUE.}

\item{group}{The variable to use to group observations.}

\item{ellipse}{Logical specifying if a normal ellipse should be drawn for
each group.}

\item{ellipse_prob}{Numerical, size of the ellipse under normal distribution.}

\item{annotations}{The variable to use to name the points.}
}
\value{
An invisible `ggplot()`.
}
\description{
Produce a PCA plot.
}
\seealso{
`ggbiplot::ggbiplot()`
}
