% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_qcs}
\alias{plot_qcs}
\title{Plot features over processing time.}
\usage{
plot_qcs(
  eset,
  features,
  ind_qcs = "Sample Identification",
  measurement_col = "Measurement Time",
  ref_index = 1
)
}
\arguments{
\item{eset}{A Biobase::ExpressionSet.}

\item{features}{Character vector of feature names for which to plot.}

\item{ind_qcs}{Name of the column to use to identify QCs (any sample
measured at least twice)}

\item{measurement_col}{The name of the column containing the measurement
time (àefaults to "Measurement Time").}

\item{ref_index}{The index of the subcolumn to use for ordering of
measurement times (typically 1, and therefore default value).}
}
\value{
An invisible `ggplot()`.
}
\description{
Plot features over processing time.
}
