% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{prop_below_LOD}
\alias{prop_below_LOD}
\title{Get the proportions of values below LOD for each metabolite.}
\usage{
prop_below_LOD(eset)
}
\arguments{
\item{eset}{A Biobase::ExpressionSet.}
}
\value{
A numeric vector of the propotions of values below LOD.
}
\description{
Get the proportions of values below LOD for each metabolite.
}
