% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_utils.R
\name{qc_rlsc}
\alias{qc_rlsc}
\title{Perform QC-RLSC for batch correction of chromatographic signal}
\usage{
qc_rlsc(
  eset,
  qc_ind = "Sample Identification",
  degree = 2,
  span = 0.75,
  measurement_col = "Measurement Time",
  ref_index = 1
)
}
\arguments{
\item{eset}{A `Biobase::ExpressionSet`.}

\item{qc_ind}{Either a vector corresponding to the `eset` samples, where `1`
indicates a QC and `2` indicates a samples. Or the name of the column from
where to derive this information. Repeatedly measured samples will be used
as QC. If more than one sample was repeatedly mesured, only the first
repeated sample identified will be used.}

\item{degree}{The degree of polynomials to be used.}

\item{span}{The parameter \ifelse{latex}{\out{$\alpha$}}{\ifelse{html}{\out{&alpha;}}{alpha}} which controls the LOESS smoothing.}

\item{measurement_col}{The name of the column from where to derive processing
order (either using the MetIDQ Mesurement Time or by a vector of integers).}

\item{ref_index}{The index of the subcolumn to use for ordering of
measurement times (typically 1, and therefore default value). Will only be
used if the column corresponding to `measurement_column` is not a vector of
integers.}
}
\value{
A `Biobase::ExpressionSet`.
}
\description{
Perform QC-RLSC for batch correction of chromatographic signal
}
\references{
Dunn et al. Nature Protocols 6, 1060-1083 (2011)
}
