% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDQ2eset.R
\name{IDQ2eset}
\alias{IDQ2eset}
\title{Transform an excel file from MetIDQ into a Biobase::ExpressionSet}
\usage{
IDQ2eset(
  path,
  col_before = "Measurement Time",
  LOD_string = "LOD \\\\(calc\\\\.\\\\) ([0-9]+)/([0-9]+) \\\\[µM\\\\]",
  LOD_rep = "\\\\1-\\\\2",
  plate_code = "Plate Bar Code"
)
}
\arguments{
\item{path}{A string saying where to find the excel file to use as source.}

\item{col_before}{String giving the name of the last column of metadata.
Typically it is called 'Measurement.Time'.}

\item{LOD_string}{String used to parse the LOD rownames.
Defaults to "LOD \\(calc\\.\\) ([0-9]+)/([0-9]+) \\[µM\\]".}

\item{LOD_rep}{String to rename the LOD columns.}

\item{plate_code}{Name of the column containing the Plate car code,
typically "Plate.Bar.Code".}
}
\value{
A Biobase::ExpressionSet object.
}
\description{
Transform an excel file from MetIDQ into a Biobase::ExpressionSet
}
