% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{msea}
\alias{msea}
\title{Perform metabolite set enrichment analysis (MSEA)}
\usage{
msea(
  metabolites,
  pathways,
  p.adj = "fdr",
  min.set.size = 5,
  fun = subramanian,
  ...
)
}
\arguments{
\item{metabolites}{A `data.frame` with the first column containing metabolite
IDs and the second column the associated P-values.}

\item{pathways}{A `data.frame` with the first column containing a pathway ID
and a second column containing a metabolite ID.}

\item{p.adj}{Method for P-value adjustment. Default: FDR. `NULL` for none.}

\item{min.set.size}{Minimal size of a set, should be at least 2.}

\item{fun}{Function to use to test for enrichment. Default is `subramanian`
to use the Kolmogorow-Smirnow test. You can provide any function taking a
vector of p-values as first argument and the indices of the
genes/metabolites of interest within these. Further arguments are allowed.}

\item{...}{Further arguments that will be passed to `fun`.}
}
\value{
A `data.frame` containing pathway ID, number of metabolites found in
  the pathway, P-value (significantly enriched) and adjusted P-value.
}
\description{
Perform metabolite set enrichment analysis (MSEA)
}
