% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{subramanian}
\alias{subramanian}
\title{Test for significance of enrichment using the Kolmogorow-Smirnow Test.}
\usage{
subramanian(p, set, nperm = 1000)
}
\arguments{
\item{p}{All P-values associated with the factor of interest .}

\item{set}{Indices of the p values associated with the pathway of interest.}

\item{nperm}{Number of permutations to get a p-value by gene permutations.}
}
\value{
A vector of p-values
}
\description{
Test for significance of enrichment using the Kolmogorow-Smirnow Test.
}
